refactor: reorganiza estrutura de módulos e corrige nomes de arquivos
- Renomeia 'carregadar_dados.py' para 'carregar_dados.py'
- Move 'visualizacao_bairros.py' para o módulo 'visualization'
- Atualiza imports nos arquivos __init__.py
- Renomeia 'cricao_mapa.py' para 'criacao.py' no módulo map_tools
