+++ quali_bus/data_analysis/__init__.py
@@ -1,5 +1,5 @@
 from .calcular_indicadores import *
-from .carregadar_dados import *
+from .carregar_dados import *
 from .analisar_dataframe import *
 from .visualizar_graficos import *
-from .visualizacao_bairros import *
+from ..visualization.visualizacao_bairros import *
